extract ocr dot wave animation
